refac: mainViewModel 만들어봤습니다..
