+++ pubspec.yaml
@@ -35,6 +35,12 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  provider: ^6.1.1
+  freezed_annotation: ^2.4.1
+  json_annotation: ^4.8.1
+  go_router: ^13.0.1
+  get_it: ^7.6.6
+  http: ^1.1.2
 
 dev_dependencies:
   flutter_test:
@@ -46,6 +52,9 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^2.0.0
+  freezed: ^2.4.6
+  json_serializable: ^6.7.1
+  build_runner: ^2.4.8
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
